requirements.txt
# Python dependencies for moments4u execution scripts
# Install with: pip install -r requirements.txt

python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
boto3>=1.34.0
celery[redis]>=5.3.0
redis>=5.0.0
insightface>=0.7.3
onnxruntime>=1.17.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
pgvector>=0.2.5
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
